requirements.txt
beautifulsoup4==4.8.2
requests==2.22.0
urllib3==1.23
Pillow==5.3.0
tqdm==4.41.1